+++ src/Oro/Bundle/ConfigBundle/Entity/Repository/ConfigRepository.php
@@ -4,6 +4,7 @@ namespace Oro\Bundle\ConfigBundle\Entity\Repository;
 
 use Doctrine\ORM\EntityRepository;
 use Oro\Bundle\ConfigBundle\Entity\Config;
+use Oro\Bundle\ConfigBundle\ORM\Hydration\ConfigObjectHydrator;
 
 /**
  * Repository class form Config entity.
@@ -18,6 +19,11 @@ class ConfigRepository extends EntityRepository
      */
     public function findByEntity($scope, $scopeId)
     {
+        $this
+            ->getEntityManager()
+            ->getConfiguration()
+            ->addCustomHydrationMode(ConfigObjectHydrator::class, ConfigObjectHydrator::class);
+
         return $this->createQueryBuilder('c')
             ->select('c, cv')
             ->leftJoin('c.values', 'cv')
@@ -25,6 +31,6 @@ class ConfigRepository extends EntityRepository
             ->setParameter('entityName', $scope)
             ->setParameter('entityId', $scopeId)
             ->getQuery()
-            ->getOneOrNullResult();
+            ->getOneOrNullResult(ConfigObjectHydrator::class);
     }
 }